Factor the loop peeling simulation into a class concerned with loops and one concerned with general-purpose propagation.
This is in preparation for using the same code to analyse calls, especially recursive ones whose inlining would boil down to inlining by another name. Relatedly, this commit also amends the type of the substituted instructions map to Map<Value*, Constant*> in order to admit Argument*, which is another child of Value but not of Instruction.