Add support for loop invariants (surprisingly difficult!) including fragments of the CFG, blocks which can be always-dead at their natural scope, and so on.
Complications include the face that killing a greater chunk of CFG means that a block which *may* be killed invariant-ly can also be killed at child scopes. The invariant annotation just cues us to check whether the outer scope knows it's already dead before trying to figure that out ourselves.

This may or may not be a time win. It should be a space win, given we'll store that information once at parent scope rather than potentially millions of times at child scope.

Also add a test that stresses the invariant logic a bit.