Fix a cornucopia of bugs, allowing a few more test cases to be emitted.
Also add a script that uses the right sequence of incantations to hammer a bitcode file into shape for processing. It uses -instcombine -jump-threading rather than -std-compile-opts to get rid of trivial instructions because some of these test cases can be completely flattened by the ordinary optimiser. It uses -loop-rotate to prepare loops for the unroller.