Add support for committing programs with resolved VFS operations
As a result the VFS tests now pass.

Also seperated the concepts of "dead" store elimination into "unused writers", which are memory operations that aren't actually going to feed any reads once integration is complete, and "dead" instructions, which are instructions which won't have any users once we're done.
This was because read instructions got confusing, as they can be found "unused" like a big memset on their buffer, and have a return value that makes them 'dead' in the eyes of the DIE code.

Note that the VFS commit code currently doesn't respect the GUI selector -- that's up next.