Introduce a caching layer for text representations of instructions.
This is because it turns out if you operator<< an Instruction you end up indexing the entire bitcode file to determine the identifiers assigned to anonymous values!

This takes a while (100ms or so for libc + coreutils), and if repeated once per debug statement (or once per instruction being written to a DOT graph) this really adds up.

To use the caching layer you surround Value*, ValCtx or MemDepResult objects with itcache(...), which constructs a wrapper with an operator<< that grabs a pass-global table of cached instructions.

Code in Save.cpp doesn't use this since its actions will alter the slot assignment.