Add a hacked up version of the postdominator code so that loops no longer assign invariant certainty to exit blocks on the strength that they must exit some time.
This basically hacks the BasicBlock succ and pred-iterators to describe an amended graph in which the loop header has no incoming edges, the latch leads to a special null node, and blocks which are neither in the loop nor have in-loop predecessors are ignored.

That turned out to be a fairly large pain however, because those iterators are supplied by GraphTraits, and GraphTraits must be specialised over something that's not BasicBlock (I introduced a wrapper for a Loop and another for a BB), and the Traits classes must be compatible with the way DominatorTreeBase<...> uses them, imitating Function objects.

This still doesn't quite work, but it's very nearly the right thing.