Add build logic for the fakestdio test.
This needs to be built as 2 modules so we can apply full optimisation to the fake stdio calls so that their store->load loop turns into a single load, a loop involving GEP bumps and a store.

This in turn is necessary because we can't at present figure out that an in-loop store-to-load dependency can be used to infer an invariant base.

Finally they can't be optimised together or the optimiser flattens the whole program.