Add support for discovering pointer base fixed points even when loads and stores are involved.
The implementation has loop forwarding in PB mode conducted within the PB solver subloop, meaning it can mimic PHI resolution by making optimistic assumptions at first and then pessimistic ones to determine if it successfully caused a recurrence.

Add a test that mimics the tricky _charpad function which relies on this property: a dynamic loop containing repeated stdio calls, each of which loads the stdio current pointer then bumps it. In order to avoid epic clobberage we need to know that *every* invocation of stdio_write in that context conserves the property that the stdio buffer pointer leads to a particular allocation.