Redesign selective optimistic solver to use a queue of pending loops and values. Each loop is solved seperately.
This is because the existing design broke when e.g. one loop fed a second, but the second happened to get explored first; it then zeroed its results and read crap from the first. This way around the second loop would be zeroed, calculate a bunch of overdefs, and do nothing. Then when the first loop gets improved it will requeue the second, which will be improved, and so on.

This is like the algorithm employed by the pessimistic solver, but on the granularity of loops (and individual instructions if they look like they're bound to a particular iteration/inlining.