Reinstate rule preventing elim of open instructions that are directly used.
As the comments state, this only really needs to be there because the resolved VC is likely invalid by the time we try to use it in commitLocalPointers. Soon I'll rewrite this to work with copies of the original functions rather than substituting in place so this kind of value reference remains valid.