Make sure to preserve otherwise dead values that are needed due to the way commit works.
Specfically, a copy of a ptr would usually be dead if all its users are resolved, but we keep it alive if the loop unroller or inliner will take a direct copy to supply an argument or header PHI, because they don't pay attention to our replacement sets.

Also improve the GUI to permit inspection of folded contexts.