Add an option for expanding calls in invariant context if their loop isn't expanded (as opposed to ignored), but for the time being revert to always behaving this way.
Also document many more syscalls, and fix a simple bug where constantexprs were accepted as resolved values.

Finally add a command-line option to supress running the loop solver when we know it'll be run outside as well. This loses the opportunity to kill blocks when invariants are established about subloops, but saves time if the subloops are nearly the same size as the outer loop.