Forbid subqueries whilst in the PB solver.
This is because there was the possibility of infinite recursion, re-entering the loop and finding the same definer repeatedly. I may try to support subqueries with backedges properly in future, possibly by retaining the visited-set in the new context.

Also fix: forbid searching for definers for const ptr null using same mechanism as filter for const-global searches.