GetElementPtrInst and related constant expressions: don't try to guess the element type
Looks like in places I had thought this was supposed to be the return type; in fact it's the type of *ptr, where ptr is the base operand. The assertion in the header file from LLVM 7+ makes clear there is little value in supplying this ourself rather than just passing null and determining it automatically.